LayoutTests/imported/w3c/web-platform-tests/speculation-rules/prefetch/no-http-cache-interference.https-expected.txt


FAIL Check that the HTTP cache doesn't cache when the server conditionally responds with 503 for prefetch requests. assert_equals: Should use the non-prefetched 200, instead of the prefetched 503 expected "200" but got "503"
